+++ examples/test/LifeTests.hs
@@ -130,7 +130,7 @@ lifeTests =
               (soa, _, irNodes, _) <- flatPrepareCore life
               js <- Ex.evaluate $ renderJS (effectfulASTFromSoA soa)
               irNodes @?= 70639
-              T.length js @?= 874063
+              T.length js @?= 874151
           , testCase "seedLiveCells stamps sparse pairs into zeroed buffers" $
               renderJS
                 ( effectfulAST